+++ mock-engine-store/src/mock_cluster/mod.rs
@@ -1,15 +1,11 @@
 // Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
 
-pub mod cluster;
 pub mod cluster_ext;
+// mod common should be private
 mod common;
 pub mod config;
-pub mod node;
-pub mod server;
-pub mod transport_simulate;
-pub mod util;
+pub mod v1;
 
-pub use cluster::*;
 pub use cluster_ext::*;
 pub use common::*;
 pub use config::{Config, MockConfig};
@@ -18,4 +14,4 @@ pub use test_raftstore::{
     new_put_cmd, new_region_leader_cmd, new_request, new_status_request, new_store,
     new_tikv_config, new_transfer_leader_cmd, sleep_ms,
 };
-pub use util::*;
+pub use v1::*;
